rename components module to ui
