+++ src/interface/star.tsx
