+++ src/pages/product/[id].tsx
@@ -1,7 +1,7 @@
-import { InfiniteScroll } from "@/components/list/infinite-scroll";
-import { Spinner } from "@/components/loading/spinner";
-import { SortDialog, useSortQuery } from "@/components/search/sort";
-import { Star } from "@/components/ui/star";
+import { InfiniteScroll } from "@/interface/list/infinite-scroll";
+import { Spinner } from "@/interface/loading/spinner";
+import { SortDialog, useSortQuery } from "@/interface/search/sort";
+import { Star } from "@/interface/star";
 import type { NextPageWithLayout } from "@/layout";
 import { Layout } from "@/layout";
 import { layoutScrollUpTracker } from "@/layout/scroll-up-tracker";
