+++ src/pages/profile/settings.tsx
@@ -2,13 +2,14 @@ import { signOut } from "@/auth";
 import type { Provider } from "@/auth/provider";
 import { providerIcons } from "@/auth/provider/icon";
 import { useCachedSession } from "@/auth/session/hooks";
-import { loadingTracker } from "@/components/loading/indicator";
-import { logToastError, toast } from "@/components/toast";
-import { Button, Input, WithLabel } from "@/components/ui";
-import { DialogOverlay } from "@/components/ui/dialog";
-import { LabeledSwitch } from "@/components/ui/switch";
 import { compressImage } from "@/image/compress";
 import { ImagePickerButton } from "@/image/image-picker";
+import { Button } from "@/interface/button";
+import { DialogOverlay } from "@/interface/dialog";
+import { Input, WithLabel } from "@/interface/input";
+import { loadingTracker } from "@/interface/loading/indicator";
+import { LabeledSwitch } from "@/interface/switch";
+import { logToastError, toast } from "@/interface/toast";
 import type { NextPageWithLayout } from "@/layout";
 import { Layout } from "@/layout";
 import { useQueryToggleState } from "@/navigation/query/hooks";
