+++ src/pages/review/[id]/edit.tsx
@@ -1,16 +1,17 @@
 import { ScrollUpButton } from "@/browser/scroll-up";
-import { InfiniteScroll } from "@/components/list/infinite-scroll";
-import { loadingTracker } from "@/components/loading/indicator";
-import { Spinner } from "@/components/loading/spinner";
-import { DebouncedSearch, useSearchQuery, useSetSearchQuery } from "@/components/search/search";
-import { logToastError, toast } from "@/components/toast";
-import { AutoresizableInput, Button, ButtonLike, Input } from "@/components/ui";
-import { DialogOverlay } from "@/components/ui/dialog";
-import { Star } from "@/components/ui/star";
-import { LabeledSwitch } from "@/components/ui/switch";
 import { useBlobUrl } from "@/image/blob";
 import { compressImage } from "@/image/compress";
 import { ImagePickerButton } from "@/image/image-picker";
+import { Button, ButtonLike } from "@/interface/button";
+import { DialogOverlay } from "@/interface/dialog";
+import { AutoresizableInput, Input } from "@/interface/input";
+import { InfiniteScroll } from "@/interface/list/infinite-scroll";
+import { loadingTracker } from "@/interface/loading/indicator";
+import { Spinner } from "@/interface/loading/spinner";
+import { DebouncedSearch, useSearchQuery, useSetSearchQuery } from "@/interface/search/search";
+import { Star } from "@/interface/star";
+import { LabeledSwitch } from "@/interface/switch";
+import { logToastError, toast } from "@/interface/toast";
 import type { NextPageWithLayout } from "@/layout";
 import { Layout } from "@/layout";
 import { getQueryParam } from "@/navigation/query";
