+++ src/trpc/infinite-query.ts
@@ -1,4 +1,4 @@
-import { logToastError } from "@/components/toast";
+import { logToastError } from "@/interface/toast";
 import type { UseTRPCInfiniteQueryResult } from "@trpc/react-query/shared";
 
 export function fetchNextPage<Q extends UseTRPCInfiniteQueryResult<unknown, unknown>>(
